tests: group every gate into tiers, run them in CI, and repair the rotted ones
The root Makefile carried twenty check targets and the tree another
dozen per-directory suites, and firmware.yml ran six of the targets and
five of the suites; the rest ran on nobody's machine. Four targets
(check-divider, check-pdp11-reference, check-elf2aout and any gate
reaching a verifier) refused to run without PYTHON in the environment,
check-exec-spool and check-cache-footprint had not linked since the swap
cursor landed in exec_subr.c (#66), the Smaller C suite executed nothing
in CI because the runner has no qemu-arm and locally faulted on the
Boot ROM float resolver reading address 0x10, and the host package's
pytest imported whichever discobsd-host the system had installed.

PYTHON defaults to python3 in share/mk/sys.mk and the root Makefile and
is exported to every sub-make. The exec spool test models the kernel's
cyclic next-fit reservation (malloc3_contiguous_next, swapnext,
swap_cursor_publish) and gains a scenario for the cursor's advance,
publication and wrap. The Smaller C execution tier links a copy of libc
with the rom_float_ members removed so libgcc's soft-float answers the
__aeabi_ helpers, refuses a link-only run under REQUIRE_QEMU=1, and
fails any link that still reaches the ROM path. pytest reads src/ ahead
of site-packages.

The root Makefile groups the gates: check-lint (shellcheck -S error over
every tracked shell script, ruff over every tracked Python file, through
tools/check-lint.sh, .shellcheckrc and ruff.toml), check-host,
check-posix-sh, check-cross, check-qemu, check-mips, check-host-package
and check-board-build (tests/rp2040/Makefile builds the on-device
programs), and check runs them all. firmware.yml runs the tiers on
Ubuntu with qemu-user, capstone, pyelftools, gcc-multilib and the
mipsel-linux-gnu cross compiler installed, and the tiers macOS can run
there. Every host test build now carries -Werror; that surfaced
sign-compare in paste.c and cksum.c, a fallthrough comment GCC does not
read in unexpand.c, and in smlrc.c three fallthroughs after the
noreturn error functions and a strlen result held in an int. The tree's
Python is at zero ruff findings. sys/arch/rp2040/doc/TESTING.md carries
the matrix; README, CLAUDE.md and DEVELOPMENT.md point at it.

+++ usr.bin/paste/paste.c
@@ -11,7 +11,7 @@ struct fdescr {
 };
 
 static void
-sequential(struct fdescr *dsc, int fdescrlen, Rune *delim, size_t delimlen)
+sequential(struct fdescr *dsc, size_t fdescrlen, Rune *delim, size_t delimlen)
 {
 	Rune c, last;
 	size_t i, d;
@@ -38,7 +38,7 @@ sequential(struct fdescr *dsc, int fdescrlen, Rune *delim, size_t delimlen)
 }
 
 static void
-parallel(struct fdescr *dsc, int fdescrlen, Rune *delim, size_t delimlen)
+parallel(struct fdescr *dsc, size_t fdescrlen, Rune *delim, size_t delimlen)
 {
 	Rune c, d;
 	size_t i, m;
@@ -120,7 +120,7 @@ main(int argc, char *argv[])
 	/* populate file list */
 	dsc = ereallocarray(NULL, argc, sizeof(*dsc));
 
-	for (i = 0; i < argc; i++) {
+	for (i = 0; i < (size_t)argc; i++) {
 		if (!strcmp(argv[i], "-")) {
 			argv[i] = "<stdin>";
 			dsc[i].fp = stdin;
@@ -136,7 +136,7 @@ main(int argc, char *argv[])
 		parallel(dsc, argc, delim_rune, delim_runelen);
 	}
 
-	for (i = 0; i < argc; i++)
+	for (i = 0; i < (size_t)argc; i++)
 		if (dsc[i].fp != stdin && fshut(dsc[i].fp, argv[i]))
 			ret |= fshut(dsc[i].fp, argv[i]);
